appleservicechallenge/requirements.txt
Flask==0.12.2
antiorm==1.2.1
asn1crypto==0.24.0
aspy.yaml==1.0.0
async-timeout==2.0.1
cached-property==1.4.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.3.4
cryptography==2.2
db==0.1.1
dill==0.2.7.1
dnspython==1.15.0
flake8==3.5.0
Flask-JSON==0.3.2
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gitdb==0.6.4
GitPython==2.0.8
graphviz==0.4.10
identify==1.0.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
multidict==4.3.1
multiprocess==0.70.5
nodeenv==1.3.0
passlib==1.7.1
pathos==0.2.1
pip==10.0.1
pluggy==0.6.0
pox==0.2.3
ppft==1.6.4.7.1
pre-commit==0.16.3
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
PyJWT==1.6.4
PyMySQL==0.8.1
pyreadline==2.1
PyYAML==3.12
requests==2.18.4
setuptools==39.2.0
six==1.11.0
smmap==0.9.0
SQLAlchemy==1.2.8
tox==2.9.1
typing==3.5.3.0
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.14.1
wheel==0.31.1
WTForms==2.2.1
yarl==1.2.0
